default ward view as an extension
