+++ packages/esm-ward-app/src/root.component.tsx
@@ -1,17 +1,20 @@
+import { ExtensionSlot, useConfig } from '@openmrs/esm-framework';
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import WardView from './ward-view/ward-view.component';
 
 const Root: React.FC = () => {
   // t('wards', 'Wards')
   const wardViewBasename = window.getOpenmrsSpaBase() + 'home/ward';
+  const { wardViewSlotName } = useConfig();
+
+  const wardView = <ExtensionSlot name={wardViewSlotName ?? 'default-ward-view-slot'} />;
 
   return (
     <main>
       <BrowserRouter basename={wardViewBasename}>
         <Routes>
-          <Route path="/" element={<WardView />} />
-          <Route path="/:locationUuid" element={<WardView />} />
+          <Route path="/" element={wardView} />
+          <Route path="/:locationUuid" element={wardView} />
         </Routes>
       </BrowserRouter>
     </main>
