+++ packages/esm-ward-app/src/ward-patient-card/default-ward/default-ward-patient-card.component.tsx
@@ -1,16 +1,20 @@
 import classNames from 'classnames';
 import React from 'react';
 import { type WardPatientCardType } from '../../types';
+import PendingItemsCardRowExtension from '../card-rows/pending-items-card-row.extension';
 import WardPatientAge from '../row-elements/ward-patient-age';
 import WardPatientBedNumber from '../row-elements/ward-patient-bed-number';
 import WardPatientIdentifier from '../row-elements/ward-patient-identifier';
 import WardPatientName from '../row-elements/ward-patient-name';
+import WardPatientTimeOnWard from '../row-elements/ward-patient-time-on-ward';
 import WardPatientCard from '../ward-patient-card.component';
 import styles from '../ward-patient-card.scss';
-import PendingItemsCardRowExtension from '../card-rows/pending-items-card-row.extension';
+import WardPatientTimeSinceAdmission from '../row-elements/ward-patient-time-since-admission';
+import AdmissionRequestNoteRowExtension from '../card-rows/admission-request-note.extension';
 
 const DefaultWardPatientCard: WardPatientCardType = (wardPatient) => {
-  const { patient, bed } = wardPatient;
+  const { patient, bed, inpatientAdmission } = wardPatient;
+  const { encounterAssigningToCurrentInpatientLocation, firstAdmissionOrTransferEncounter } = inpatientAdmission ?? {};
 
   const card = (
     <WardPatientCard
@@ -20,9 +24,13 @@ const DefaultWardPatientCard: WardPatientCardType = (wardPatient) => {
           <WardPatientName patient={patient} />
           <WardPatientIdentifier patient={patient} />
           <WardPatientAge patient={patient} />
+          <WardPatientTimeSinceAdmission firstAdmissionOrTransferEncounter={firstAdmissionOrTransferEncounter} />
+          <WardPatientTimeOnWard
+            encounterAssigningToCurrentInpatientLocation={encounterAssigningToCurrentInpatientLocation}
+          />
         </div>
       }
-      rows={[<PendingItemsCardRowExtension {...wardPatient} />]}
+      rows={[<PendingItemsCardRowExtension {...wardPatient} />, <AdmissionRequestNoteRowExtension {...wardPatient} />]}
       wardPatient={wardPatient}
     />
   );
