+++ packages/esm-ward-app/src/ward-view-header/ward-view-header.component.tsx
@@ -1,19 +1,21 @@
-import React from 'react';
+import React, { type ReactNode } from 'react';
 import styles from './ward-view-header.scss';
 import AdmissionRequestsBar from './admission-requests-bar.component';
 import useWardLocation from '../hooks/useWardLocation';
 import WardMetrics from './ward-metrics.component';
 
-interface WardViewHeaderProps {}
+interface WardViewHeaderProps {
+  wardPendingPatients: ReactNode;
+}
 
-const WardViewHeader: React.FC<WardViewHeaderProps> = () => {
+const WardViewHeader: React.FC<WardViewHeaderProps> = ({ wardPendingPatients }) => {
   const { location } = useWardLocation();
 
   return (
     <div className={styles.wardViewHeader}>
       <h4>{location?.display}</h4>
       <WardMetrics />
-      <AdmissionRequestsBar />
+      <AdmissionRequestsBar {...{ wardPendingPatients }} />
     </div>
   );
 };
