+++ packages/esm-ward-app/src/ward-workspace/admission-request-workspace/admission-requests.workspace.tsx
@@ -1,21 +1,16 @@
-import React from 'react';
+import React, { type ReactNode } from 'react';
 import styles from './admission-requests-workspace.scss';
 import AdmissionRequestCard from '../admission-request-card/admission-request-card.component';
 import { Search } from '@carbon/react';
-import { ErrorState } from '@openmrs/esm-framework';
+import { type DefaultWorkspaceProps, ErrorState } from '@openmrs/esm-framework';
 import { useTranslation } from 'react-i18next';
 import { useInpatientRequest } from '../../hooks/useInpatientRequest';
 import { type InpatientRequest } from '../../types';
 
-interface AdmissionRequestsWorkspaceProps {}
-const AdmissionRequestsWorkspace: React.FC<AdmissionRequestsWorkspaceProps> = () => {
-  // note: useAppContext() does not work here for some reason, so we call `useInpatientRequest`
-  // directly. See: https://openmrs.atlassian.net/browse/O3-4020
-  const {
-    inpatientRequests,
-    isLoading: isLoadingInpatientRequests,
-    error: errorFetchingInpatientRequests,
-  } = useInpatientRequest(['ADMIT', 'TRANSFER']);
+interface AdmissionRequestsWorkspaceProps extends DefaultWorkspaceProps {
+  wardPendingPatients: ReactNode;
+}
+const AdmissionRequestsWorkspace: React.FC<AdmissionRequestsWorkspaceProps> = ({ wardPendingPatients }) => {
   const { t } = useTranslation();
   const [searchTerm, setSearchTerm] = React.useState('');
   const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -33,29 +28,7 @@ const AdmissionRequestsWorkspace: React.FC<AdmissionRequestsWorkspaceProps> = ()
         disabled
       />
 
-      <div className={styles.content}>
-        {isLoadingInpatientRequests ? (
-          <>Loading</>
-        ) : errorFetchingInpatientRequests ? (
-          <ErrorState
-            headerTitle={t('admissionRequests', 'Admission requests')}
-            error={errorFetchingInpatientRequests}
-          />
-        ) : (
-          <>
-            {inpatientRequests.map((request: InpatientRequest, i) => (
-              <AdmissionRequestCard
-                key={`admission-request-card-${i}`}
-                patient={request.patient}
-                visit={request.visit}
-                bed={null}
-                inpatientRequest={request}
-                inpatientAdmission={null}
-              />
-            ))}
-          </>
-        )}
-      </div>
+      <div className={styles.content}>{wardPendingPatients}</div>
     </div>
   );
 };
